backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==1.10.13
psutil==5.9.8
chromadb==0.4.24
python-dotenv==1.0.1
httpx==0.26.0
requests==2.31.0
Pillow==10.2.0
GitPython==3.1.42
pygit2==1.13.1
